// frontend/bindings/narria/backend/models/models.ts
// Cynhyrchwyd y ffeil hon yn awtomatig. PEIDIWCH Â MODIWL
// This file is automatically generated. DO NOT EDIT

// eslint-disable-next-line @typescript-eslint/ban-ts-comment
// @ts-ignore: Unused imports
import { Create as $Create } from "@wailsio/runtime";

// eslint-disable-next-line @typescript-eslint/ban-ts-comment
// @ts-ignore: Unused imports
import * as uuid$0 from "../../../github.com/google/uuid/models.js";

export class UserData {
    "uuid": uuid$0.UUID;
    "username": string;
    "password": string;
    "recovery": string;

    /** Creates a new UserData instance. */
    constructor($$source: Partial<UserData> = {}) {
        if (!("uuid" in $$source)) {
            this["uuid"] = "";
        }
        if (!("username" in $$source)) {
            this["username"] = "";
        }
        if (!("password" in $$source)) {
            this["password"] = "";
        }
        if (!("recovery" in $$source)) {
            this["recovery"] = "";
        }

        Object.assign(this, $$source);
    }

    /**
     * Creates a new UserData instance from a string or object.
     */
    static createFrom($$source: any = {}): UserData {
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        return new UserData($$parsedSource as Partial<UserData>);
    }
}